refactor(tts): move WordTTS into shared examples
