+++ docs/how-to/configure-tts.md
@@ -134,7 +134,7 @@ Set `synthesis_mode` on the catalog entry (hydrated as `tts_synthesis_mode`). Ma
 
 ### Word-level input streaming and timestamps
 
-All examples use Pipecat's `NvidiaTTSService` by default, which keeps Magpie Multilingual, Magpie Zeroshot, and Chatterbox switchable through the service catalog. For Magpie TTS Multilingual 1.10.0 or newer, [`NvidiaWordTTSService`](../../src/nvidia_word_tts.py) is an optional drop-in subclass that adds word-level input streaming and timestamp-based LLM context commits. It requires `nvidia-riva-client` 2.27.0 or newer.
+All examples use Pipecat's `NvidiaTTSService` by default, which keeps Magpie Multilingual, Magpie Zeroshot, and Chatterbox switchable through the service catalog. For Magpie TTS Multilingual 1.10.0 or newer, [`NvidiaWordTTSService`](../../src/examples/shared/nvidia_word_tts.py) is an optional drop-in subclass that adds word-level input streaming and timestamp-based LLM context commits. It requires `nvidia-riva-client` 2.27.0 or newer.
 
 To opt in for a custom example, change only the service import and constructor:
 
@@ -145,7 +145,7 @@ from pipecat.services.nvidia.tts import NvidiaTTSService
 tts = NvidiaTTSService(**tts_kwargs)
 
 # Opt in to Magpie 1.10.0+ word streaming and timestamp commits
-from nvidia_word_tts import NvidiaWordTTSService
+from examples.shared.nvidia_word_tts import NvidiaWordTTSService
 
 tts = NvidiaWordTTSService(**tts_kwargs)
 ```
